requirements.txt
Django>=5.0,<6.0
django-ninja>=1.1.0
django-ninja-jwt>=5.3.0
psycopg2-binary>=2.9.9
channels>=4.1.0
channels-redis>=4.2.0
redis>=5.0.0
daphne>=4.1.2
django-cors-headers>=4.4.0
python-dotenv>=1.0.1
